part of service implementation and tests  added

+++ service/src/main/java/by/eventcat/EventServiceImpl.java
@@ -1,7 +1,10 @@
 package by.eventcat;
 
 
+import by.eventcat.custom.exceptions.CustomErrorCodes;
+import by.eventcat.custom.exceptions.ServiceException;
 import org.springframework.dao.DataAccessException;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -22,37 +25,80 @@ public class EventServiceImpl implements EventService{
     }
 
     @Override
-    public List<Event> getAllEvents() throws DataAccessException {
-        return eventDao.getAllEvents();
+    public List<Event> getAllEvents() throws DataAccessException, ServiceException {
+        List<Event> events = eventDao.getAllEvents();
+        if (events == null) throw new ServiceException(CustomErrorCodes.NO_CALLING_DATA_FOUND);
+        return events;
     }
 
     @Override
-    public List<Event> getAllEventsByEventPlaceName(String eventPlaceName) throws DataAccessException {
-        return eventDao.getAllEventsByEventPlaceName(eventPlaceName);
+    public List<Event> getAllEventsByEventPlaceName(String eventPlaceName) throws DataAccessException, ServiceException {
+        if(eventPlaceName == null || eventPlaceName.length() < 2){
+            throw new ServiceException(CustomErrorCodes.INCORRECT_INDEX);
+        }
+        List<Event> events;
+        try{
+            events = eventDao.getAllEventsByEventPlaceName(eventPlaceName);
+        } catch (EmptyResultDataAccessException ex){
+            throw new ServiceException(CustomErrorCodes.NO_CALLING_DATA_FOUND);
+        }
+        return events;
     }
 
     @Override
-    public List<Event> getAllEventsByCategoryId(Category category) throws DataAccessException {
-        return eventDao.getAllEventsByCategoryId(category);
+    public List<Event> getAllEventsByCategoryId(Category category) throws DataAccessException, ServiceException {
+        if (category.getCategoryId() <= 0) throw new ServiceException(CustomErrorCodes.INCORRECT_INDEX);
+        List<Event> events;
+        try{
+            events = eventDao.getAllEventsByCategoryId(category);
+        } catch (EmptyResultDataAccessException ex){
+            throw new ServiceException(CustomErrorCodes.NO_CALLING_DATA_FOUND);
+        }
+        return events;
     }
 
     @Override
-    public Event getEventById(Integer eventId) throws DataAccessException {
-        return eventDao.getEventById(eventId);
+    public Event getEventById(Integer eventId) throws DataAccessException, ServiceException {
+        if (eventId <= 0) throw new ServiceException(CustomErrorCodes.INCORRECT_INDEX);
+        Event event = eventDao.getEventById(eventId);
+        if (event == null)  throw new ServiceException(CustomErrorCodes.NO_CALLING_DATA_FOUND);
+        return event;
     }
 
     @Override
-    public Integer addEvent(Event event) throws DataAccessException {
+    public Integer addEvent(Event event) throws DataAccessException, ServiceException {
+        if (event.getCategory().getCategoryId() <= 0 ||
+                event.getEventName() == null ||
+                event.getEventName().length() < 2 ||
+                event.getEventPlace() == null ||
+                event.getEventPlace().length() < 2){
+            throw new ServiceException(CustomErrorCodes.INCORRECT_INPUT_DATA);
+        }
         return eventDao.addEvent(event);
     }
 
     @Override
-    public int updateEvent(Event event) throws DataAccessException {
-        return eventDao.updateEvent(event);
+    public int updateEvent(Event event) throws DataAccessException, ServiceException {
+        if (event.getEventId() <= 0) throw new ServiceException(CustomErrorCodes.INCORRECT_INDEX);
+        if (event.getCategory().getCategoryId() <= 0 ||
+                event.getEventName() == null ||
+                event.getEventName().length() < 2 ||
+                event.getEventPlace() == null ||
+                event.getEventPlace().length() < 2){
+            throw new ServiceException(CustomErrorCodes.INCORRECT_INPUT_DATA);
+        }
+        int rowsAffected = eventDao.updateEvent(event);
+        if (rowsAffected == 0) throw new ServiceException(CustomErrorCodes.NO_ACTIONS_MADE);
+        if (rowsAffected > 1) throw new ServiceException(CustomErrorCodes.ACTIONS_ERROR);
+        return rowsAffected;
     }
 
     @Override
-    public int deleteEvent(Integer eventId) throws DataAccessException {
-        return eventDao.deleteEvent(eventId);
+    public int deleteEvent(Integer eventId) throws DataAccessException, ServiceException {
+        if (eventId <= 0) throw new ServiceException(CustomErrorCodes.INCORRECT_INDEX);
+        int rowsAffected = eventDao.deleteEvent(eventId);
+        if (rowsAffected == 0) throw new ServiceException(CustomErrorCodes.NO_ACTIONS_MADE);
+        if (rowsAffected > 1) throw new ServiceException(CustomErrorCodes.ACTIONS_ERROR);
+        return rowsAffected;
     }
 }
